First commit: PassthroughtHandsFreeController project
This is the first commit of the project, which intends to implement a
hands free controller to an eletric wheelchair using a meta quest 3
headset.

On this first version, the controller is based on the interaction of the
center eye anchor and the scene mesh generated automatically by the OS
after a enviroment setup procedure. User is already able to look to a
position on the floor throught the passthrough layer and recieve visual
feedback on the mixed reality.