First version of UI and new Log wrapper
This commit adds a initial implementation of the UI that will be show to
the user when they stare on the desired position on the ground where
they want the weelchair to go. The canvas design is based on  Meta prefabs
provided on the Alll-in-one SDK. It still lacking the functionallity
thats enables the user to select its buttons using just the camera
direction.

Also, I've created a Debug wrapper class to enable more comprehensive
(and pretty) logging. It is a static class, so every component on the
project can access its functions without needing to have a reference to
it.