Movement UI improved and functionating
This commits adds functionality to the MovementCanvas UI added on the
last commit. Now, the user is able to select both buttons by pointing the
laser pointer at one of them for 2 seconds using just the center eye camera.

The UIHelpers prefab was based on Meta's implementation available on the
PassthroughApi samples. The code was changed in a way that the
OVRInputeModule of the EventSystem uses the center eye camera transform as its
ray transform, making the UI interaction rays being cast in the
direction of the user sight.

I've also deleted the CameraGazeUICursor object and scripts because all
its pretended functionallity could be implemented using just the
MovementCanvas and the already existent CameraGazerCursor.

Morover, I've removed the OVROverlayCanvas component from the MovementCanvas
because it was making the laser pointer casted by the UIHelpers to be
rendered bellow the canvas backgroung. I pretend to make it work in the
future, but the interface is working just fine without it by now.