Wheelchair tracking using controllers as anchors
asiuhdashdiak ash dahskd alkjs lkajsh dans kadna loisuhd kans kdhas
Meta's support for QR code tracking is an unreliable, experimental
feature. Alternative QR tracking methods require passthrough camera
access, which is unavailable over Meta Link. This limitation makes
development slow, as each minor change would necessitate a full
application rebuild. Consequently, this approach abandons QR code
tracking in favor of using the controllers as spatial anchors to
determine the wheelchair's pose.

The idea is to 3D-print two controller stands and attach them to
the wheelchair's armrests. This ensures the controllers remain in
the scene and are tracked by the Quest's operating system, even if
the user cannot hold them. This method provides the wheelchair's
pose in every frame, enabling the application of movement control
logic.

In addition to this new tracking method, this commit introduces a
simple RestThru API to communicate with the wheelchair's server and
implement a simple point-and-go function.